# LiteWeightApiDocumentation/public/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LiteWeight API",
    "version": "v1"
  },
  "paths": {
...
    "/self": {
      "get": {
        "tags": [
          "Self"
        ],
        "summary": "Get Self",
        "description": "Returns the user that is currently authenticated.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Self"
        ],
        "summary": "Create Self",
        "description": "Creates a user in the database using the email/firebase id in the authenticated token.\r\n<br /><br />Note that a single verified, authenticated user can only exist once in the database - this is determined by the firebase UUID in the authenticated token.<br/><br/><div style=\"background-color: #2f3031;padding: 12px 0px 12px 12px;border-left: 5px solid #F90258;\">Potential 400 Errors: <b>AlreadyExists, InvalidRequest</b></div>",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSelfRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSelfRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSelfRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Self"
        ],
        "summary": "Delete Self",
        "description": "Deletes a user and all data associated with it.<br /><br />\r\nData deleted: user, workouts belonging to user, images belonging to user. The user is also removed as a friend from any other users, and any friend requests they sent are canceled.<br/><br/><div style=\"background-color: #2f3031;padding: 12px 0px 12px 12px;border-left: 5px solid #02F88F;\"><b>Sends Push Notification</b></div>",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/self/profile-picture": {
      "put": {
        "tags": [
          "Self"
        ],
        "summary": "Update Profile Picture",
        "description": "Updates the user's profile picture. Note that it simply replaces the old picture using the same image url.<br/><br/><div style=\"background-color: #2f3031;padding: 12px 0px 12px 12px;border-left: 5px solid #F90258;\">Potential 400 Errors: <b>InvalidRequest</b></div>",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfilePictureRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfilePictureRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfilePictureRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/self/set-firebase-messaging-token": {
      "put": {
        "tags": [
          "Self"
        ],
        "summary": "Set Firebase Messaging Token",
        "description": "Sets the firebase messaging token to the authenticated user. This enables the authenticated user's ability to receive push notifications, or removes it if the token is null.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetFirebaseMessagingTokenRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetFirebaseMessagingTokenRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetFirebaseMessagingTokenRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/self/all-friend-requests-seen": {
      "put": {
        "tags": [
          "Self"
        ],
        "summary": "Set All Friend Requests Seen",
        "description": "Sets all friend requests on the authenticated user as seen.",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/self/settings": {
      "put": {
        "tags": [
          "Self"
        ],
        "summary": "Set Settings",
        "description": "Sets the settings on the authenticated user.<br/><br/><div style=\"background-color: #2f3031;padding: 12px 0px 12px 12px;border-left: 5px solid #F90258;\">Potential 400 Errors: <b>InvalidRequest</b></div>",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsResponse"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsResponse"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsResponse"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/self/current-workout": {
      "put": {
        "tags": [
          "Self"
        ],
        "summary": "Set Current Workout",
        "description": "Sets the current workout of the authenticated user to the specified workout, if it exists.<br/><br/><div style=\"background-color: #2f3031;padding: 12px 0px 12px 12px;border-left: 5px solid #F90258;\">Potential 400 Errors: <b>WorkoutNotFound</b></div>",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentWorkoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentWorkoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/self/received-workouts/all-seen": {
      "put": {
        "tags": [
          "Self"
        ],
        "summary": "Set All Received Workouts Seen",
        "description": "Sets all received workouts on the authenticated user as seen.",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/self/received-workouts/{sharedWorkoutId}/seen": {
      "put": {
        "tags": [
          "Self"
        ],
        "summary": "Set Received Workout Seen",
        "description": "Sets a given received workout on the authenticated user as seen.",
        "parameters": [
          {
            "name": "sharedWorkoutId",
            "in": "path",
            "description": "Received workout to set as seen",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadRequestResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message describing the error.",
            "nullable": true,
            "example": "An example error message"
          },
          "errorType": {
            "type": "string",
            "description": "Specific type of error.",
            "nullable": true,
            "example": "InvalidRequest"
          },
          "requestErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelBindingError"
            },
            "description": "Lists any errors associated when binding the request.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateSelfRequest": {
        "required": [
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string",
            "description": "Username of the new user - must be unique. Note that the username will be converted to lowercase.",
            "example": "randy_bo_bandy"
          },
          "profilePictureData": {
            "type": "string",
            "description": "Base 64 encoding of the image to upload. If not specified, then the user will have a default profile picture.",
            "format": "byte",
            "nullable": true,
            "example": "iVBORw0KGgoAAAANSUhEUgAAAlgAAAJ"
          },
          "metricUnits": {
            "type": "boolean",
            "description": "Should the created user have metric units enabled?"
          }
        },
        "additionalProperties": false
      },
...
      "ForbiddenResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "Forbidden access"
          }
        },
        "additionalProperties": false
      },
...
      "ModelBindingError": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "description": "Property that is causing the error.",
            "nullable": true,
            "example": "name"
          },
          "message": {
            "type": "string",
            "description": "Error message for problematic property.",
            "nullable": true,
            "example": "The name property cannot be empty"
          }
        },
        "additionalProperties": false
      },
...
      "ResourceNotFoundResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "The workout with id=cfbb1fd4-6d29-465c-9ea0-4938407ac65b was not found"
          }
        },
        "additionalProperties": false
      },
...
      "ServerErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "The server has encountered an error"
          }
        },
        "additionalProperties": false
      },
      "SetCurrentWorkoutRequest": {
        "type": "object",
        "properties": {
          "currentWorkoutId": {
            "type": "string",
            "description": "Id of the workout to set as the current workout. Null signifies no workout is selected as current.",
            "nullable": true,
            "example": "4d17e3e7-edf6-41c0-ade2-fdd624b5c9bb"
          }
        },
        "additionalProperties": false
      },
...
      "SetFirebaseMessagingTokenRequest": {
        "type": "object",
        "properties": {
          "firebaseMessagingToken": {
            "type": "string",
            "description": "Firebase token to associate with the user. Enables push notifications to be sent to the device associated with the token. Null removes the ability for the user to receive notifications.",
            "nullable": true,
            "example": "c_-_jNKuQI-9USNk0c9uEY:APA91bG53pjv4wNcg7m2h2d1yfCjVfidEj7AXyIu8ddNYz2_Stwy0J7znQayzltXDgEL8Q9tSj8i2yx8hQiSuNDKlZtGm3cEmJTlLMupmJVTc1g_LdMPdx9VOL2hC3vBTseIcmBDYjRG"
          }
        },
        "additionalProperties": false
      },
...
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "Unauthorized access"
          }
        },
        "additionalProperties": false
      },
      "UpdateProfilePictureRequest": {
        "required": [
          "profilePictureData"
        ],
        "type": "object",
        "properties": {
          "profilePictureData": {
            "type": "string",
            "description": "Base 64 encoding of the image to upload.",
            "format": "byte",
            "example": "iVBORw0KGgoAAAANSUhEUgAAAlgAAAJ"
          }
        },
        "additionalProperties": false
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the user",
            "nullable": true,
            "example": "f1e03cd1-e62c-4a53-84ed-498c72776fc2"
          },
          "username": {
            "type": "string",
            "description": "Username of the user.",
            "nullable": true,
            "example": "barbell_bill"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user.",
            "nullable": true,
            "example": "barbellB12@gmail.com"
          },
          "profilePicture": {
            "type": "string",
            "description": "File path of the user's profile picture.",
            "nullable": true,
            "example": "0f1d96c3-ca22-4657-9f9b-136bb4621985.jpg"
          },
          "firebaseMessagingToken": {
            "type": "string",
            "description": "Firebase token registered for the user's device to receive push notifications.",
            "nullable": true,
            "example": "c-Id_13FTDf5c5PZGjsmg9:APA91bFG4w3Qnd16l0WDZzFYco3_71-_X595oncjilyKhSJBZC9FUD4mjg2lE68HCcDQR80Y6GDTuz-zbXK69V9D-jTBs4aXhv7LBAUTGzoC8h91q_QV-H1vz6XUGF8Mcob84_2-izMI"
          },
          "premiumToken": {
            "type": "string",
            "description": "Token indicating the user has purchased LiteWeight premium. Currently not used.",
            "nullable": true,
            "example": "a704f441-8ee3-471b-ac8a-abb0b7d8249a"
          },
          "currentWorkoutId": {
            "type": "string",
            "description": "Workout Id that the user is currently on. Null signifies the user is not working on a workout.",
            "nullable": true,
            "example": "b209f062-36fa-4089-aca0-31df4815744f"
          },
          "workoutsSent": {
            "type": "integer",
            "description": "Total number of workouts sent.",
            "format": "int32",
            "example": 14
          },
          "settings": {
            "$ref": "#/components/schemas/UserSettingsResponse"
          },
          "workouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutInfoResponse"
            },
            "description": "List of workouts the user owns.",
            "nullable": true
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnedExerciseResponse"
            },
            "description": "List of exercises the user owns.",
            "nullable": true
          },
          "friends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FriendResponse"
            },
            "description": "List of friends for the user.",
            "nullable": true
          },
          "friendRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FriendRequestResponse"
            },
            "description": "List of pending friend requests for the user.",
            "nullable": true
          },
          "receivedWorkouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedWorkoutInfoResponse"
            },
            "description": "List of workouts sent to the user.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserSettingsResponse": {
        "type": "object",
        "properties": {
          "privateAccount": {
            "type": "boolean",
            "description": "If true, the user will not receive any friend requests or shared workouts from users who they are not friends with."
          },
          "updateDefaultWeightOnSave": {
            "type": "boolean",
            "description": "If true, the default value of exercises will be updated when saving a workout if the completed exercise's weight is greater than the default weight.",
            "example": true
          },
          "updateDefaultWeightOnRestart": {
            "type": "boolean",
            "description": "If true, the default value of exercises will be updated when restarting a workout if the completed exercise's weight is greater than the default weight.",
            "example": true
          },
          "metricUnits": {
            "type": "boolean",
            "description": "If true, display weights as metric (only used on the UI - values returned from this API are always in imperial units)."
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "BearerDefinition": {
        "type": "http",
        "description": "Token authentication. \n\n 'Bearer TOKEN'\n\nBearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}